// experiments/active/multi-tenant-backend/src/payload-types.ts
export interface Tenant {
  id: number;
  name: string;
  /**
   * Used for domain-based tenant handling
   */
  domain?: string | null;
  /**
   * Used for url paths, example: /tenant-slug/page-slug
   */
  slug: string;
  /**
   * If checked, logging in is not required to read. Useful for building public pages.
   */
  allowPublicRead?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface User {
  id: number;
  password?: string | null;
  role?: ('super-admin' | 'user') | null;
  username?: string | null;
  /**
   * Applies to your current tenant.
   */
  tenantRole?: ('tenant-admin' | 'tenant-viewer') | null;
  tenants?:
    | {
        tenant: number | Tenant;
        role: 'tenant-admin' | 'tenant-viewer';
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
}